Refactor the only .cpp file to pure C
